ui: remove use of transient buffers
Transient buffers are destroyed and recreated at every frame.
The default size of a transient buffer (`init.limits.transientVbSize`) is
rather large and this causes the destruction, recreation and upload of
large GPU buffers.

A dynamic buffer is used instead which is recreated only when the ui needs
more space for vertices and indices. This prevents the recreation of the
buffers and leaves the uploads to only the size of buffers needed to be
done every frame.